优化请求和响应消息
